refactor: split ModelsPage into tab components (927→425 lines)
Extract LocalModelsTab, LibraryTab, CloudTab, and EngineTab from the
monolithic ModelsPage.vue. Each tab now owns its state and logic;
ModelsPage becomes an orchestrating shell of ~425 lines.
